refactoring streamlit_app
